--- test.txt
@@ -1,4 +1,0 @@
-fsdfsdfsdfsdfsdfdsfdf
-sfsdfsdfsdfsqweqwe
-sfsdgsdgsdfewg3et
-dsfdgsgsdagsdgds
